Organize config and rm tasks
